+++ src/math.cpp
@@ -130,8 +130,8 @@ double logChoose(int n, int k) {
 }
 
 void printMtx(const vector<vector<int>> &a) {
-  for (int i = 0; i < a.size(); ++i) {
-    for (int j = 0; j < a[i].size(); ++j) {
+  for (size_t i = 0; i < a.size(); ++i) {
+    for (size_t j = 0; j < a[i].size(); ++j) {
       cout << a[i][j] << ' ';
     }
     cout << "\n";
